Remove unused isLoaded property from metabolites store.
Actions which require metabolites previously used store.getTotalCount() > 0 to determine if the job has metabolites.
But when store is filtered the totalCount is the number of metabolites passing filter not the total number of metabolites in the job.
Added totalUnfiltered to metabolites store which is the total unfiltered count of metabolites.