Ignore Invalid if during rpc.add_structures also structure_database is given. If structure_database params is given and ms data is available do an annotate sub-command with the structure_database.
The annotate options from the form will be ignored because of presents of row in run table of jobs results.db.
For example when ionisation mode is 1 in run table while annotation should be done with -1 then then 1 will be used and no candidates will be retrieved, because it uses the wrong mz.