Added delete button to job grid on workspace page (fixes #50). Implemented it as DELETE verb on /magma/results/<jobid>.
Moved set of job description from results page to workspace page.
Implemented it as PUT verb on /magma/results/<jobid>.
Also allow MS filename to be altered.
Edit by clicking cell, pressing enter or focus on something else will save to database on server.

Had issues with saving description and then ms_filename on the same job.
It was causing 'database is locked' exception.
The issue did not appear in unittest so a functional test was made which duplicated the exception.
After much searching it was caused because a SELECT was performed within an UPDATE.
Preventing the SELECT by caching SELECT results before UPDATE is started fixes issue.