Add launcher_url property to Job class and job table in user db. On job submission the launcher_url is filled with the job url which is returned by the job launcher.
The status page has bee ExtJS-ified, so a cancel button could be added.
The cancel button shows a confirmation dialog before sending a DELETE request to the server.
The status page uses meta refresh to update the status page.
The refreshing could happen during the cancelling, to prevent this the meta refresh was replaced with a progressbar which polls the job status via a ajax call each interval.

Replace usage of urllib2 library by requests library.
For cancelling of a job a HTTP request has to be done with DELETE HTTP method.
To do this using urllib2 requires you to jump to hopes, by extending request class.
The requests library has much easier support for the DELETE HTTP method.

Refs #187.