Added transaction manager to job db.
Using the app in a threaded wsgi servers like waitress and uwsgi caused errors like:
ProgrammingError: SQLite objects created in a thread can only be used in that same thread.The object was created in thread id 139710103127808 and this is thread id 139710220900160

By using a transaction manager the database connections are not reused between threads. So no more errors.